requirements.txt
gpiozero==2.0
hyperlpr3==0.1.3
numpy==1.24.2
opencv_python==4.8.1.78
paho_mqtt==1.6.1
picamera2==0.3.16
Pillow==10.1.0
RPi.GPIO==0.7.1a4
smbus==1.1.post2
